GUI single camera looking decent
