+++ sessions/5_cameras/config.toml
@@ -4,7 +4,11 @@ CreationDate = 2023-02-07T08:31:15.128437
 port = 0
 resolution = [ 1280, 720,]
 rotation_count = 0
+error = 0.444
+camera_matrix = [ [ "741.156450733475", "0.0", "625.8681130786296",], [ "0.0", "734.043039807219", "372.50416072475235",], [ "0.0", "0.0", "1.0",],]
+distortion = [ [ "0.05452425785658533", "-0.030389650991086665", "0.005734953618762244", "5.3103846111261454e-05", "-0.18412320048155464",],]
 exposure = -6
+grid_count = 24
 ignore = false
 verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 
@@ -12,11 +16,11 @@ verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 port = 1
 resolution = [ 1280, 720,]
 rotation_count = 0
-error = 0.64
-camera_matrix = [ [ "1465.1751089825566", "0.0", "683.6405940226846",], [ "0.0", "1449.2468238889232", "397.34160035273237",], [ "0.0", "0.0", "1.0",],]
-distortion = [ [ "-0.042171612511387215", "0.34731048867509273", "-0.0028571476699573253", "0.011094818774842908", "-1.3282427481356034",],]
+error = 0.58
+camera_matrix = [ [ "1257.4519792700291", "0.0", "660.911526874092",], [ "0.0", "1271.6018488863265", "273.4696546240674",], [ "0.0", "0.0", "1.0",],]
+distortion = [ [ "0.041272201345104724", "-0.4014831608231816", "0.0034321775950695733", "-0.004674359963819289", "1.227263787020917",],]
 exposure = -3
-grid_count = 22
+grid_count = 20
 ignore = false
 verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 
@@ -24,11 +28,11 @@ verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 port = 2
 resolution = [ 1280, 720,]
 rotation_count = 0
-error = 0.519
-camera_matrix = [ [ "645.9800059334519", "0.0", "647.4765381895428",], [ "0.0", "642.7707250670561", "389.6206495664435",], [ "0.0", "0.0", "1.0",],]
-distortion = [ [ "0.17412298010487934", "-0.2482221426294052", "0.0029313721952783033", "-0.00179043074233148", "0.12620814139938288",],]
+error = 0.605
+camera_matrix = [ [ "657.0590943651505", "0.0", "636.805365188361",], [ "0.0", "660.2554955389062", "405.79423683958095",], [ "0.0", "0.0", "1.0",],]
+distortion = [ [ "0.19353565100888573", "-0.15714424387438963", "0.006540933964275408", "-0.005664835841041637", "-0.10528723445449427",],]
 exposure = -4
-grid_count = 44
+grid_count = 20
 ignore = false
 verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 
@@ -36,11 +40,11 @@ verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 port = 3
 resolution = [ 1280, 720,]
 rotation_count = 0
-error = 0.587
-camera_matrix = [ [ "615.4210534200573", "0.0", "652.2912514100597",], [ "0.0", "607.5821167299271", "316.0196744453672",], [ "0.0", "0.0", "1.0",],]
-distortion = [ [ "0.16884126574559488", "-0.25094201114293385", "-0.005383869917316663", "0.0028954822431506576", "0.12316671737891444",],]
-exposure = -4
-grid_count = 38
+error = 0.452
+camera_matrix = [ [ "621.2478419371772", "0.0", "652.9246542984728",], [ "0.0", "621.511606436301", "324.0427603869688",], [ "0.0", "0.0", "1.0",],]
+distortion = [ [ "0.1616061542879378", "-0.21043421448215577", "6.714132581925542e-05", "-0.0004812784177563003", "0.08626165617811646",],]
+exposure = -6
+grid_count = 30
 ignore = false
 verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 
@@ -48,11 +52,11 @@ verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
 port = 4
 resolution = [ 1280, 720,]
 rotation_count = 0
-error = 0.3
-camera_matrix = [ [ "988.8847653560879", "0.0", "638.7987591413655",], [ "0.0", "990.439219027749", "392.01160065996214",], [ "0.0", "0.0", "1.0",],]
-distortion = [ [ "-0.38145325504102784", "0.07101277530426597", "-0.011545217575620503", "0.0029254164932602704", "0.04265025406878891",],]
-exposure = -7
-grid_count = 45
+error = 0.383
+camera_matrix = [ [ "890.0250288769084", "0.0", "665.4283399262999",], [ "0.0", "885.0622827253966", "368.79365603804644",], [ "0.0", "0.0", "1.0",],]
+distortion = [ [ "-0.3654636391895171", "0.4644458857209638", "3.661635311566276e-05", "-0.009431356696076315", "-0.81202271680149",],]
+exposure = -8
+grid_count = 20
 ignore = false
 verified_resolutions = [ [ 640, 480,], [ 1280, 720,],]
