+++ sessions/laptop/config.toml
@@ -2,9 +2,13 @@ CreationDate = 2022-12-22T10:43:16.354534
 
 [cam_0]
 port = 0
-resolution = [ 1280, 720,]
+resolution = [ 1920, 1080,]
 rotation_count = 0
+error = 0.25
+camera_matrix = [ [ "1133.262824247143", "0.0", "939.0217779409916",], [ "0.0", "1141.757102361169", "546.326824359205",], [ "0.0", "0.0", "1.0",],]
+distortion = [ [ "0.09433579314328489", "-0.1753966668972666", "0.0022121166236822227", "0.001071059546039067", "-0.02666020168998023",],]
 exposure = -7
+grid_count = 21
 ignore = false
 verified_resolutions = [ [ 640, 480,], [ 1280, 720,], [ 1920, 1080,],]
